common: fix warnings and some shutdown errors

+++ byteps/common/cpu_reducer.cc
@@ -15,7 +15,6 @@
 
 #include "cpu_reducer.h"
 #include "global.h"
-#include "logging.h"
 
 namespace byteps {
 namespace common {
@@ -63,7 +62,7 @@ int CpuReducer::_sum_float32(void* dst, void* src, size_t len) {
     auto d = (float*)dst;
     auto s = (float*)src;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (size_t i = 0; i < len / 4; ++i) {
+    for (size_t i = 0; i < len / (size_t) 4; ++i) {
         d[i] = d[i] + s[i];
     }
     return 0;
@@ -73,7 +72,7 @@ int CpuReducer::_sum_float64(void* dst, void* src, size_t len) {
     auto d = (double*)dst;
     auto s = (double*)src;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (size_t i = 0; i < len / 8; ++i) {
+    for (size_t i = 0; i < len / (size_t) 8; ++i) {
         d[i] = d[i] + s[i];
     }
     return 0;
@@ -84,7 +83,7 @@ int CpuReducer::_sum_float16(void* dst, void* src, size_t len) {
     auto d = (float*)dst;
     auto s = (float*)src;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (size_t i = 0; i < len / 4; ++i) {
+    for (size_t i = 0; i < len / (size_t) 4; ++i) {
         d[i] = d[i] + s[i];
     }
     return 0;
@@ -104,7 +103,7 @@ int CpuReducer::_sum_int32(void* dst, void* src, size_t len) {
     auto d = (int*)dst;
     auto s = (int*)src;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (size_t i = 0; i < len / 4; ++i) {
+    for (size_t i = 0; i < len / (size_t) 4; ++i) {
         d[i] = d[i] + s[i];
     }
     return 0;
@@ -124,7 +123,7 @@ int CpuReducer::_sum_int64(void* dst, void* src, size_t len) {
     auto d = (long long*)dst;
     auto s = (long long*)src;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (size_t i = 0; i < len / 8; ++i) {
+    for (size_t i = 0; i < len / (size_t) 8; ++i) {
         d[i] = d[i] + s[i];
     }
     return 0;
@@ -157,7 +156,7 @@ int CpuReducer::_sum_float32(void* dst, void* src1, void* src2, size_t len) {
     auto s1 = (float*)src1;
     auto s2 = (float*)src2;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (int i = 0; i < len / 4; ++i) {
+    for (int i = 0; i < len / (size_t) 4; ++i) {
         d[i] = s1[i] + s2[i];
     }
     return 0;
@@ -168,7 +167,7 @@ int CpuReducer::_sum_float64(void* dst, void* src1, void* src2, size_t len) {
     auto s1 = (double*)src1;
     auto s2 = (double*)src2;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (int i = 0; i < len / 8; ++i) {
+    for (int i = 0; i < len / (size_t) 8; ++i) {
         d[i] = s1[i] + s2[i];
     }
     return 0;
@@ -180,7 +179,7 @@ int CpuReducer::_sum_float16(void* dst, void* src1, void* src2, size_t len) {
     auto s1 = (float*)src1;
     auto s2 = (float*)src2;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (int i = 0; i < len / 4; ++i) {
+    for (int i = 0; i < len / (size_t) 4; ++i) {
         d[i] = s1[i] + s2[i];
     }
     return 0;
@@ -202,7 +201,7 @@ int CpuReducer::_sum_int32(void* dst, void* src1, void* src2, size_t len) {
     auto s1 = (int*)src1;
     auto s2 = (int*)src2;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (int i = 0; i < len / 4; ++i) {
+    for (int i = 0; i < len / (size_t) 4; ++i) {
         d[i] = s1[i] + s2[i];
     }
     return 0;
@@ -224,7 +223,7 @@ int CpuReducer::_sum_int64(void* dst, void* src1, void* src2, size_t len) {
     auto s1 = (long long*)src1;
     auto s2 = (long long*)src2;
 #pragma omp parallel for simd num_threads(_num_threads)
-    for (int i = 0; i < len / 8; ++i) {
+    for (int i = 0; i < len / (size_t) 8; ++i) {
         d[i] = s1[i] + s2[i];
     }
     return 0;
